Fix crash when disabling plugin
If the widget is the toggle button in the main window (t1), this means
that the toggle button for the plugin option widget (t2) will have been
destroyed at this point. We only need to check if we clicked t2 widget
and set t1 to the same value in this case, since t1 is always valid.