Support WAYFIRE_PLUGIN_XML_PATH
This is being added to Wayfire to allow plugins to be installed
outside of Wayfire's prefix.  In the same spirit, Wayfire and wf-shell
may be installed to different prefixes, and this patch accomodates
that as well.

I also got rid of the intermediate wayfire_share_dir variable in
meson.build because its name was confusing now that there is
wayfire_metadata_dir and wayfire_sysconf_dir, where the wayfire_
prefix means something else.