Fix plugin toggling bug
When toggling a plugin whose name is a substring of another plugin in the
list, it was getting confused about when and what to add or remove. Here
we make the logic a bit smarter to make sure the plugin name is isolated
by spaces or the beginning/end matches the beginning/end of the plugins
string. Thanks to @nwg-piotr for finding this.