Fix ordering of combo boxes (#53)
When iterating over a std::map, the order is not guaranteed. In this case,
it was ordering the items by alphabetic order of name, instead of the order
in the xml files. This was causing the initially selected item to not match
the actual value in some cases. Fix this by using a std::vector of std::pairs.