Fix int and animation option writes (#86)
Trying to set int or animate options does not work when typing them into
the spin button field. Scrolling and clicking the tick boxes to change the
value does, however. Apparently, this is a problem with gtkmm when calling
get_value_as_int() in the changed handler. It does not affect double options,
which use get_value(). This solution uses the C API to work around the problem.