Make wcm single instance
We have to set org.gtk. prefix for the application identifier or the app can
be started multiple times. This commit also moves from .png to .svg icon for
wcm. We install wcm.svg to two different locations so it can be found by apps
like wf-panel and still have wcm find it where it expects.